requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alabaster==0.7.8
appdirs==1.4.3
# apturl==0.5.2
astroid==2.3.3
async-timeout==5.0.1
attrs==19.3.0
Babel==2.6.0
backcall==0.1.0
bleach==3.1.1
blinker==1.4
# Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.4.1
click==7.1.2
cloudpickle==1.3.0
# cmake==3.31.4
colorama==0.4.3
# command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
# cupshelpers==1.0
cycler==0.12.1
datasets==3.1.0
dbus-python==1.2.16
decorator==4.4.2
# defer==1.0.6
defusedxml==0.6.0
dill==0.3.8
distlib==0.3.0
distro==1.4.0
# distro-info==0.23+ubuntu1.1
docutils==0.16
entrypoints==0.3
filelock==3.0.12
fonttools==4.55.5
frozenlist==1.5.0
fsspec==2024.9.0
gpytorch==1.13
greenlet==0.4.15
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.27.1
idna==2.8
imagesize==1.2.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.4
jaxtyping==0.2.19
jedi==0.15.2
Jinja2==2.10.1
joblib==1.4.2
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==18.0.1
kiwisolver==1.4.7
# language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
linear-operator==0.5.3
lit==18.1.8
logilab-common==1.4.3
# louis==3.12.0
macaroonbakery==1.3.1
MarkupSafe==1.1.0
matplotlib==3.7.5
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==4.2.0
mpmath==1.3.0
msgpack==0.6.2
multidict==6.1.0
multiprocess==0.70.16
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
numpydoc==0.7.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-htop==1.2.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
packaging==24.2
pandas==2.0.3
pandocfilters==1.4.2
parso==0.5.2
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
propcache==0.2.0
protobuf==5.29.3
psutil==5.5.1
pyarrow==17.0.0
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyparsing==2.4.6
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.15.5
pysvn==1.9.9
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
pytz==2024.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QtAwesome==0.4.4
qtconsole==4.6.0
QtPy==1.9.0
ranger-fm==1.9.3
regex==2024.11.6
reportlab==3.5.34
requests==2.32.3
requests-unixsocket==0.2.0
roman==2.0.0
rope==0.10.5
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentencepiece==0.2.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
Sphinx==1.8.5
spyder==3.3.6
spyder-kernels==0.5.2
ssh-import-id==5.10
swag==2.0.5
sympy==1.13.3
systemd-python==234
tabulate==0.9.0
termcolor==2.4.0
testpath==0.4.4
threadpoolctl==3.5.0
timm==1.0.14
tokenizers==0.20.3
torch==2.0.1
torchvision==0.15.2
tornado==5.1.1
tqdm==4.67.1
traitlets==4.3.3
transformers==4.46.3
triton==2.0.0
typeguard==4.4.0
typer==0.3.2
typing-extensions==4.12.2
tzdata==2024.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
xkit==0.0.0
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
